requirements.txt
asgiref==3.8.1
attrs==25.1.0
backports.zoneinfo==0.2.1
Django==4.2.19
django-cors-headers==4.4.0
django-jazzmin==3.0.1
django-nested-admin==4.1.1
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.2.1
importlib-resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
pkgutil-resolve-name==1.3.10
python-monkey-business==1.1.0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
sqlparse==0.5.3
typing-extensions==4.12.2
uritemplate==4.1.1
zipp==3.20.2
